+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* Added Solstice event challenges to the Progress page.
+
 ## 7.26.0 <span class="changelog-date">(2022-07-17)</span>
 
 * Worked around a Bungie.net API bug where Vanguard reset count was reported under Strange Favor (Dares of Eternity) instead.